Add Email API, integrate into fixture and docs
Introduce a new Email sub-API for sending (SMTP via nodemailer) and receiving (IMAP via imapflow) emails in tests. Adds src/interactions/Email.ts with send/receive/receiveAll/clean and MIME extraction/saving of HTML for browser navigation; extends enum/Options with email types and interfaces. Integrates email into ElementInteractions and Steps (constructor + configureEmail), exposes BaseFixtureOptions in the test fixture (baseFixture accepts emailCredentials), and re-exports from src/index.ts. Update package.json to add imapflow, nodemailer and @types/nodemailer, and wire CI workflows to pass email-related secrets to test and publish jobs. Documentation (README, CLAUDE.md, skills) added usage examples and API docs, and a set of new tests/specs were added (old unit-tests.spec.ts removed).
