+++ .github/workflows/publish.yml
@@ -37,7 +37,13 @@ jobs:
           docker compose up --build -d
 
       - name: 🧪  Run Tests
-        run: npm test  
+        run: npm test
+        env:
+          SENDER_EMAIL: ${{ secrets.SENDER_EMAIL }}
+          SENDER_PASSWORD: ${{ secrets.SENDER_PASSWORD }}
+          SENDER_SMTP_HOST: ${{ secrets.SENDER_SMTP_HOST }}
+          RECEIVER_EMAIL: ${{ secrets.RECEIVER_EMAIL }}
+          RECEIVER_PASSWORD: ${{ secrets.RECEIVER_PASSWORD }}
 
       - name: 📦  Publish Package
         run: npm publish --access public --provenance
