+++ .github/workflows/test.yml
@@ -38,6 +38,12 @@ jobs:
 
       - name: 🧪 Run Tests
         run: npm test
+        env:
+          SENDER_EMAIL: ${{ secrets.SENDER_EMAIL }}
+          SENDER_PASSWORD: ${{ secrets.SENDER_PASSWORD }}
+          SENDER_SMTP_HOST: ${{ secrets.SENDER_SMTP_HOST }}
+          RECEIVER_EMAIL: ${{ secrets.RECEIVER_EMAIL }}
+          RECEIVER_PASSWORD: ${{ secrets.RECEIVER_PASSWORD }}
 
       - name: 📝 Check Types
         run: npx tsc --noEmit
