+++ package.json
@@ -29,6 +29,7 @@
     "@playwright/test": "^1.58.2",
     "@types/debug": "^4.1.13",
     "@types/node": "^20.0.0",
+    "@types/nodemailer": "^7.0.11",
     "pw-element-repository": "^0.0.4",
     "typescript": "^5.0.0"
   },
@@ -53,6 +54,8 @@
   "author": "Umut Ay Bora",
   "license": "MIT",
   "dependencies": {
-    "debug": "^4.4.3"
+    "debug": "^4.4.3",
+    "imapflow": "^1.2.16",
+    "nodemailer": "^8.0.3"
   }
 }
