+++ src/utils/ElementUtilities.ts
@@ -1,11 +1,11 @@
 import { Locator } from '@playwright/test';
+import { log } from '../logger/Logger';
 
 /**
  * Utility class to handle standardized waiting logic across the framework.
  */
 export class Utils {
     private readonly timeout: number;
-
     /**
      * @param timeout - Optional timeout in milliseconds. Defaults to 30000.
      */
@@ -41,11 +41,11 @@ export class Utils {
             if (message.includes('strict mode violation')) {
                 console.warn(`Locator resolved to multiple elements. Waiting on first element instead.`);
                 try { await locator.first().waitFor({ state, timeout: this.timeout }); }
-                catch { console.warn(`First element failed to reach state '${state}' within ${this.timeout}ms. Proceeding...`); }
+                catch { log.warn(`First element failed to reach state '${state}' within ${this.timeout}ms...`); }
                 return;
             }
 
-            console.warn(`Element failed to reach state '${state}' within ${this.timeout}ms. Proceeding...`);
+            log.warn(`Element failed to reach state '${state}' within ${this.timeout}ms...`);
         }
     }
 }
