--- tests/unit-tests.spec.ts
@@ -1,2642 +1,0 @@
-import { test, expect } from './fixture/StepFixture';
-import { ElementRepository } from 'pw-element-repository';
-import { Steps } from '../src/steps/CommonSteps';
-import { DropdownSelectType, ScreenshotOptions } from '../src/enum/Options';
-import { createLogger } from '../src/logger/Logger';
-
-const log = createLogger('tests');
-
-test.describe('E2E Facade Implementation Suite', () => {
-
-  test('TC_001: Complete Form Submission (Core API)', async ({ page, repo, steps, interactions, contextStore }) => {
-
-    await test.step('Navigate to the website', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('Verify Category Count', async () => {
-      await steps.verifyCount('HomePage', 'categories', { exactly: 8 });
-    });
-
-    await test.step('Open Forms Page and verify navigation', async () => {
-      await steps.click('HomePage', 'formsCard');
-      await steps.verifyAbsence('HomePage', 'categories');
-    });
-
-    await test.step('Verify Page Title', async () => {
-      await steps.verifyUrlContains('/forms');
-      await steps.verifyText('FormsPage', 'title', 'Submission Form');
-    });
-
-    await test.step('Fill Standard Inputs', async () => {
-      contextStore.put('Name', 'Automated Tester');
-      contextStore.put('Email', 'AutomatedTester@email.com');
-      contextStore.put('Mobile', '0000000000');
-      contextStore.put('Current Address', 'Prinsenstraat, 1015 DB Amsterdam');
-
-      await steps.fill('FormsPage', 'nameInput', contextStore.get('Name'));
-      await steps.fill('FormsPage', 'emailInput', contextStore.get('Email'));
-      await steps.fill('FormsPage', 'mobileInput', contextStore.get('Mobile'));
-      await steps.fill('FormsPage', 'addressInput', contextStore.get('Current Address'));
-    });
-
-    await test.step('Select a Random Enabled Gender', async () => {
-      const gender = await steps.selectDropdown('FormsPage', 'genderDropdown', {
-        type: DropdownSelectType.RANDOM
-      });
-      contextStore.put('Gender', gender);
-    });
-
-    await test.step('Handle Date Picker and Data Extraction', async () => {
-      await steps.click('FormsPage', 'dateOfBirthInput');
-      await steps.waitForState('FormsPage', 'todayCell', 'visible');
-      await steps.verifyPresence('FormsPage', 'todayCell');
-      await steps.click('FormsPage', 'todayCell');
-
-      await steps.verifyPresence('FormsPage', 'datePickerSubmitButton');
-      await steps.click('FormsPage', 'datePickerSubmitButton');
-
-      const now = new Date();
-      const dobValue = `${now.getFullYear()}-${now.getMonth() + 1}-${now.getDate()}`;
-      contextStore.put('Date of Birth', dobValue);
-
-      await steps.click('FormsPage', 'hobbiesInput');
-    });
-
-    await test.step('Submit Form and Verify Modal via verifyListedElement', async () => {
-      await steps.click('FormsPage', 'submitButton');
-      await steps.verifyPresence('FormsPage', 'table');
-
-      for (const [key, expectedValue] of contextStore.entries()) {
-        await steps.verifyListedElement('FormsPage', 'submissionRows', {
-          text: key,
-          child: { pageName: 'FormsPage', elementName: 'submissionValue' },
-          expectedText: expectedValue
-        });
-      }
-    });
-
-    log('TC_001 Complete Form Submission — passed');
-  });
-
-  test('TC_002: Drag and Drop Interactions', async ({ page, steps, repo, interactions }) => {
-
-    await test.step('Navigate to Sortable page via homepage', async () => {
-      await steps.navigateTo('/');
-      await steps.click('HomePage', 'interactionsCard');
-      await steps.verifyUrlContains('/sortable');
-    });
-
-    await test.step('Drag Item A to the Second List', async () => {
-      const dropZone = await repo.getByText(page, 'SortablePage', 'dropZones', 'Second List');
-
-      await steps.dragAndDropListedElement('SortablePage', 'sortableItems', 'Item A', { target: dropZone! });
-
-      await interactions.verify.textContains(dropZone!, 'Item A');
-    });
-
-    log('TC_002 Drag and Drop Interactions — passed');
-  });
-
-  test('TC_003: Negative Assertions - Expecting Verifications to Fail', async ({ page, repo }) => {
-
-    const steps = new Steps(page, repo, 1000); // Shorten timeout for negative assertions
-
-    await test.step('Navigate to the website', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('verifyAbsence on a visible element should throw', async () => {
-      let errorCaught = false;
-      try {
-        await steps.verifyAbsence('HomePage', 'categories');
-      } catch (error) {
-        errorCaught = true;
-        log('Caught expected error: verifyAbsence failed correctly');
-      }
-      expect(errorCaught).toBeTruthy();
-    });
-
-    await test.step('verifyCount with an incorrect number should throw', async () => {
-      let errorCaught = false;
-      try {
-        await steps.verifyCount('HomePage', 'categories', { exactly: 99 });
-      } catch (error) {
-        errorCaught = true;
-        log('Caught expected error: verifyCount failed correctly');
-      }
-      expect(errorCaught).toBeTruthy();
-    });
-
-    log('TC_003 Negative Assertions — passed');
-  });
-
-  test('TC_004: Wait For State - Warning behavior on incorrect state', async ({ page, repo }) => {
-    const steps = new Steps(page, repo, 500);
-
-    await test.step('Navigate to the website', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('waitForState should swallow the error and log a warning', async () => {
-      let errorCaught = false;
-
-      log('Intentionally waiting for a timeout to trigger the warning mechanism...');
-      try {
-        await steps.waitForState('HomePage', 'categories', 'hidden');
-      } catch (error) {
-        errorCaught = true;
-      }
-
-      expect(errorCaught).toBeFalsy();
-      log('waitForState safely swallowed the timeout error and proceeded');
-    });
-
-    log('TC_004 Wait For State Warning Behavior — passed');
-  });
-
-  test('TC_005: Click Random - Category Navigation', async ({ steps }) => {
-
-    await test.step('Navigate to the website', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('Click a random category and verify navigation', async () => {
-      await steps.clickRandom('HomePage', 'categories');
-      await steps.verifyAbsence('HomePage', 'categories');
-    });
-
-    log('TC_005 Click Random — passed');
-  });
-
-  test('TC_006: Verify Count - greaterThan and lessThan', async ({ page, repo }) => {
-    const steps = new Steps(page, repo, 3000);
-
-    await test.step('Navigate to the website', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('verifyCount with greaterThan (positive)', async () => {
-      await steps.verifyCount('HomePage', 'categories', { greaterThan: 3 });
-    });
-
-    await test.step('verifyCount with lessThan (positive)', async () => {
-      await steps.verifyCount('HomePage', 'categories', { lessThan: 10 });
-    });
-
-    await test.step('verifyCount with greaterThan polls until timeout (negative)', async () => {
-      const start = Date.now();
-      let errorCaught = false;
-      try {
-        await steps.verifyCount('HomePage', 'categories', { greaterThan: 8 });
-      } catch {
-        errorCaught = true;
-      }
-      const elapsed = Date.now() - start;
-      expect(errorCaught).toBeTruthy();
-      expect(elapsed).toBeGreaterThan(2500);
-      log('greaterThan polling confirmed: timed out after %dms', elapsed);
-    });
-
-    log('TC_006 Verify Count greaterThan/lessThan — passed');
-  });
-
-});
-
-test.describe('TC_007: verifyState - All Playwright element states', () => {
-
-  test('positive state assertions', async ({ page, repo, steps, interactions }) => {
-
-    await test.step('Navigate to Forms page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('HomePage', 'formsCard');
-      await steps.verifyUrlContains('/forms');
-    });
-
-    await test.step('visible: title is visible', async () => {
-      await steps.verifyState('FormsPage', 'title', 'visible');
-    });
-
-    await test.step('attached: title is attached to the DOM', async () => {
-      await steps.verifyState('FormsPage', 'title', 'attached');
-    });
-
-    await test.step('inViewport: title is in viewport', async () => {
-      await steps.verifyState('FormsPage', 'title', 'inViewport');
-    });
-
-    await test.step('enabled: submit button is enabled', async () => {
-      await steps.verifyState('FormsPage', 'submitButton', 'enabled');
-    });
-
-    await test.step('editable: name input is editable', async () => {
-      await steps.verifyState('FormsPage', 'nameInput', 'editable');
-    });
-
-    await test.step('focused: name input is focused after clicking', async () => {
-      await steps.click('FormsPage', 'nameInput');
-      await steps.verifyState('FormsPage', 'nameInput', 'focused');
-    });
-
-    await test.step('Navigate to Radio Buttons page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('HomePage', 'elementsCard');
-      await steps.verifyUrlContains('/radiobuttons');
-    });
-
-    await test.step('disabled: the No radio button is disabled', async () => {
-      await steps.verifyState('RadioButtonsPage', 'disabledRadio', 'disabled');
-    });
-
-    await test.step('checked: Yes radio is checked after clicking', async () => {
-      await steps.click('RadioButtonsPage', 'yesRadio');
-      await steps.verifyState('RadioButtonsPage', 'yesRadio', 'checked');
-    });
-
-    await test.step('hidden: FormsPage title is hidden on a different page', async () => {
-      await steps.verifyState('FormsPage', 'title', 'hidden', 500);
-    });
-
-    log('TC_007 verifyState — passed');
-  });
-});
-
-test.describe('TC_008: navigateTo resolves relative URLs via Playwright baseURL', () => {
-  test.use({ baseURL: 'https://umutayb.github.io/vue-test-app/' });
-
-  test('navigates with a relative URL', async ({ steps }) => {
-    await test.step('Navigate using a relative URL', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('Verify the home page loaded', async () => {
-      await steps.verifyCount('HomePage', 'categories', { exactly: 8 });
-    });
-
-    await test.step('verifyUrlContains escapes regex metacharacters', async () => {
-      await steps.verifyUrlContains('vue-test-app/');
-    });
-
-    log('TC_008 navigateTo relative URL — passed');
-  });
-});
-
-// ==========================================
-// Phase 3: Elements Category Tests
-// ==========================================
-
-test.describe('TC_009: Buttons Page', () => {
-
-  test('button clicks update result text', async ({ steps }) => {
-
-    await test.step('Navigate to Buttons page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-    });
-
-    await test.step('Click Primary button and verify result', async () => {
-      await steps.click('ButtonsPage', 'primaryButton');
-      await steps.verifyTextContains('ButtonsPage', 'resultText', 'Primary');
-    });
-
-    await test.step('Click Secondary button and verify result', async () => {
-      await steps.click('ButtonsPage', 'secondaryButton');
-      await steps.verifyTextContains('ButtonsPage', 'resultText', 'Secondary');
-    });
-
-    await test.step('Click Danger button and verify result', async () => {
-      await steps.click('ButtonsPage', 'dangerButton');
-      await steps.verifyTextContains('ButtonsPage', 'resultText', 'Danger');
-    });
-
-    await test.step('Verify disabled button is disabled', async () => {
-      await steps.verifyState('ButtonsPage', 'disabledButton', 'disabled');
-    });
-
-    await test.step('Verify loading button is disabled', async () => {
-      await steps.verifyState('ButtonsPage', 'loadingButton', 'disabled');
-    });
-
-    await test.step('Verify size variants are visible', async () => {
-      await steps.verifyPresence('ButtonsPage', 'smallButton');
-      await steps.verifyPresence('ButtonsPage', 'mediumButton');
-      await steps.verifyPresence('ButtonsPage', 'largeButton');
-    });
-
-    log('TC_009 Buttons Page — passed');
-  });
-});
-
-test.describe('TC_010: Text Inputs Page', () => {
-
-  test('fill inputs and verify values display', async ({ steps }) => {
-
-    await test.step('Navigate to Text Inputs page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('Fill text input and verify', async () => {
-      await steps.fill('TextInputsPage', 'textInput', 'hello world');
-      await steps.verifyInputValue('TextInputsPage', 'textInput', 'hello world');
-    });
-
-    await test.step('Fill email input and verify', async () => {
-      await steps.fill('TextInputsPage', 'emailInput', 'test@example.com');
-      await steps.verifyInputValue('TextInputsPage', 'emailInput', 'test@example.com');
-    });
-
-    await test.step('Fill number input and verify', async () => {
-      await steps.fill('TextInputsPage', 'numberInput', '42');
-      await steps.verifyInputValue('TextInputsPage', 'numberInput', '42');
-    });
-
-    await test.step('Fill textarea and verify', async () => {
-      await steps.fill('TextInputsPage', 'textareaInput', 'multi-line text');
-      await steps.verifyInputValue('TextInputsPage', 'textareaInput', 'multi-line text');
-    });
-
-    await test.step('Verify disabled input is disabled', async () => {
-      await steps.verifyState('TextInputsPage', 'disabledInput', 'disabled');
-    });
-
-    await test.step('Verify values display updates', async () => {
-      await steps.verifyText('TextInputsPage', 'valuesDisplay', undefined, { notEmpty: true });
-    });
-
-    await test.step('Type sequentially in text input', async () => {
-      await steps.fill('TextInputsPage', 'textInput', '');
-      await steps.typeSequentially('TextInputsPage', 'textInput', 'typed');
-      await steps.verifyInputValue('TextInputsPage', 'textInput', 'typed');
-    });
-
-    log('TC_010 Text Inputs Page — passed');
-  });
-});
-
-test.describe('TC_011: Checkboxes & Toggles Page', () => {
-
-  test('check, uncheck, and toggle interactions', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to Checkboxes page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'checkboxesLink');
-      await steps.verifyUrlContains('/checkboxes');
-    });
-
-    await test.step('Check the unchecked checkbox', async () => {
-      await steps.check('CheckboxesPage', 'uncheckedCheckbox');
-      await steps.verifyState('CheckboxesPage', 'uncheckedCheckbox', 'checked');
-    });
-
-    await test.step('Uncheck the checked checkbox', async () => {
-      await steps.uncheck('CheckboxesPage', 'checkedCheckbox');
-    });
-
-    await test.step('Verify disabled checkbox is disabled', async () => {
-      await steps.verifyState('CheckboxesPage', 'disabledCheckbox', 'disabled');
-    });
-
-    await test.step('Verify disabled-checked checkbox is both checked and disabled', async () => {
-      await steps.verifyState('CheckboxesPage', 'disabledCheckedCheckbox', 'checked');
-      await steps.verifyState('CheckboxesPage', 'disabledCheckedCheckbox', 'disabled');
-    });
-
-    await test.step('Toggle switches via label click (hidden inputs)', async () => {
-      // Toggle inputs are display:none; click the parent <label> instead
-      const toggleOffLabel = page.locator(repo.getSelector('CheckboxesPage', 'toggleOff')).locator('..');
-      const toggleOnLabel = page.locator(repo.getSelector('CheckboxesPage', 'toggleOn')).locator('..');
-      await toggleOffLabel.click();
-      await toggleOnLabel.click();
-    });
-
-    await test.step('Verify state summary updates', async () => {
-      await steps.verifyText('CheckboxesPage', 'stateSummary', undefined, { notEmpty: true });
-    });
-
-    log('TC_011 Checkboxes & Toggles — passed');
-  });
-});
-
-test.describe('TC_012: Sliders Page', () => {
-
-  test('set slider values and verify display', async ({ steps }) => {
-
-    await test.step('Navigate to Sliders page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'slidersLink');
-      await steps.verifyUrlContains('/sliders');
-    });
-
-    await test.step('Set basic slider to 75 and verify value', async () => {
-      await steps.setSliderValue('SlidersPage', 'basicSlider', 75);
-      await steps.verifyTextContains('SlidersPage', 'basicSliderValue', '75');
-    });
-
-    await test.step('Set stepped slider to 50 and verify value', async () => {
-      await steps.setSliderValue('SlidersPage', 'steppedSlider', 50);
-      await steps.verifyTextContains('SlidersPage', 'steppedSliderValue', '50');
-    });
-
-    await test.step('Verify disabled slider is disabled', async () => {
-      await steps.verifyState('SlidersPage', 'disabledSlider', 'disabled');
-    });
-
-    await test.step('Verify range slider values display', async () => {
-      await steps.verifyText('SlidersPage', 'rangeValue', undefined, { notEmpty: true });
-    });
-
-    log('TC_012 Sliders Page — passed');
-  });
-});
-
-test.describe('TC_013: Drag Progress Page', () => {
-
-  test('drag progress controls and preset buttons', async ({ steps }) => {
-
-    await test.step('Navigate to Drag Progress page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'dragProgressLink');
-      await steps.verifyUrlContains('/drag-progress');
-    });
-
-    await test.step('Click Set 50% button and verify', async () => {
-      await steps.click('DragProgressPage', 'setHalfButton');
-      await steps.verifyTextContains('DragProgressPage', 'progressValue', '50');
-    });
-
-    await test.step('Click Set 100% button and verify', async () => {
-      await steps.click('DragProgressPage', 'setFullButton');
-      await steps.verifyTextContains('DragProgressPage', 'progressValue', '100');
-    });
-
-    await test.step('Click Reset button and verify', async () => {
-      await steps.click('DragProgressPage', 'resetButton');
-      await steps.verifyTextContains('DragProgressPage', 'progressValue', '0');
-    });
-
-    await test.step('Verify progress track and handle are present', async () => {
-      await steps.verifyPresence('DragProgressPage', 'progressTrack');
-      await steps.verifyPresence('DragProgressPage', 'progressHandle');
-    });
-
-    log('TC_013 Drag Progress Page — passed');
-  });
-});
-
-// ==========================================
-// Phase 3: Forms Category Tests
-// ==========================================
-
-test.describe('TC_014: Dropdown Page', () => {
-
-  test('native single select and custom dropdown', async ({ steps }) => {
-
-    await test.step('Navigate to Dropdown page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'dropdownLink');
-      await steps.verifyUrlContains('/dropdown');
-    });
-
-    await test.step('Select a random option from the single select', async () => {
-      const selected = await steps.selectDropdown('DropdownSelectPage', 'singleSelect');
-      expect(selected).toBeTruthy();
-    });
-
-    await test.step('Verify single select value is displayed', async () => {
-      await steps.verifyText('DropdownSelectPage', 'singleValue', undefined, { notEmpty: true });
-    });
-
-    await test.step('Select by value from single select', async () => {
-      await steps.selectDropdown('DropdownSelectPage', 'singleSelect', {
-        type: DropdownSelectType.VALUE,
-        value: 'Canada'
-      });
-      await steps.verifyTextContains('DropdownSelectPage', 'singleValue', 'Canada');
-    });
-
-    await test.step('Open custom dropdown and verify list appears', async () => {
-      await steps.click('DropdownSelectPage', 'customDropdownButton');
-      await steps.verifyPresence('DropdownSelectPage', 'customDropdownList');
-    });
-
-    log('TC_014 Dropdown Page — passed');
-  });
-});
-
-test.describe('TC_015: File Upload Page', () => {
-
-  test('single file upload', async ({ steps }) => {
-
-    await test.step('Navigate to File Upload page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'fileUploadLink');
-      await steps.verifyUrlContains('/file-upload');
-    });
-
-    await test.step('Upload a single file and verify filename displayed', async () => {
-      await steps.uploadFile('FileUploadPage', 'singleFileInput', 'tests/fixtures/test-upload.txt');
-      await steps.verifyTextContains('FileUploadPage', 'singleFileName', 'test-upload.txt');
-    });
-
-    log('TC_015 File Upload Page — passed');
-  });
-});
-
-test.describe('TC_016: Autocomplete Page', () => {
-
-  test('type to filter and select suggestion', async ({ page, steps }) => {
-
-    await test.step('Navigate to Autocomplete page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'autocompleteLink');
-      await steps.verifyUrlContains('/autocomplete');
-    });
-
-    await test.step('Type "Uni" to trigger suggestions', async () => {
-      await steps.typeSequentially('AutocompletePage', 'searchInput', 'Uni', 50);
-    });
-
-    await test.step('Click "United States" suggestion', async () => {
-      const suggestion = page.locator('li').filter({ hasText: 'United States' }).first();
-      await suggestion.click();
-    });
-
-    await test.step('Verify selected value', async () => {
-      await steps.verifyTextContains('AutocompletePage', 'selectedValue', 'United States');
-    });
-
-    await test.step('Clear and verify reset', async () => {
-      await steps.click('AutocompletePage', 'clearButton');
-      await steps.verifyInputValue('AutocompletePage', 'searchInput', '');
-    });
-
-    log('TC_016 Autocomplete Page — passed');
-  });
-});
-
-// ==========================================
-// Phase 3: Alerts, Frame & Windows Category Tests
-// ==========================================
-
-test.describe('TC_017: Alerts Page - Click Types', () => {
-
-  test('click, right-click, and double-click trigger native alerts', async ({ page, steps }) => {
-
-    await test.step('Navigate to Alerts page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'alertsLink');
-      await steps.verifyUrlContains('/alerts');
-    });
-
-    await test.step('Click Me triggers alert with "Single click!"', async () => {
-      let dialogMessage = '';
-      page.once('dialog', async (dialog) => {
-        dialogMessage = dialog.message();
-        await dialog.accept();
-      });
-      await steps.click('AlertsPage', 'clickMeButton');
-      expect(dialogMessage).toBe('Single click!');
-    });
-
-    await test.step('Right Click Me triggers alert with "Right click!"', async () => {
-      let dialogMessage = '';
-      page.once('dialog', async (dialog) => {
-        dialogMessage = dialog.message();
-        await dialog.accept();
-      });
-      await steps.rightClick('AlertsPage', 'rightClickButton');
-      expect(dialogMessage).toBe('Right click!');
-    });
-
-    await test.step('Double Click Me triggers alert with "Double click!"', async () => {
-      let dialogMessage = '';
-      page.once('dialog', async (dialog) => {
-        dialogMessage = dialog.message();
-        await dialog.accept();
-      });
-      await steps.doubleClick('AlertsPage', 'doubleClickButton');
-      expect(dialogMessage).toBe('Double click!');
-    });
-
-    log('TC_017 Alerts Page Click Types — passed');
-  });
-});
-
-test.describe('TC_018: Alerts Page - New Tab', () => {
-
-  test('new tab opens and can be closed', async ({ page, steps }) => {
-
-    await test.step('Navigate to Alerts page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'alertsLink');
-      await steps.verifyUrlContains('/alerts');
-    });
-
-    await test.step('Click New Tab and switch to it', async () => {
-      const newPage = await steps.switchToNewTab(async () => {
-        await steps.click('AlertsPage', 'newTabButton');
-      });
-      await steps.verifyTabCount(2);
-      await newPage.close();
-    });
-
-    await test.step('Verify back to single tab', async () => {
-      await steps.verifyTabCount(1);
-    });
-
-    log('TC_018 Alerts Page New Tab — passed');
-  });
-});
-
-test.describe('TC_019: Modal Page', () => {
-
-  test('open modal, confirm, and verify status', async ({ steps }) => {
-
-    await test.step('Navigate to Modal page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'modalLink');
-      await steps.verifyUrlContains('/modal');
-    });
-
-    await test.step('Open modal and verify overlay appears', async () => {
-      await steps.click('ModalPage', 'openButton');
-      await steps.verifyPresence('ModalPage', 'overlay');
-    });
-
-    await test.step('Click Confirm and verify status', async () => {
-      await steps.click('ModalPage', 'confirmButton');
-      await steps.verifyAbsence('ModalPage', 'overlay');
-      await steps.verifyTextContains('ModalPage', 'status', 'confirmed');
-    });
-
-    await test.step('Reopen modal and cancel', async () => {
-      await steps.click('ModalPage', 'openButton');
-      await steps.verifyPresence('ModalPage', 'overlay');
-      await steps.click('ModalPage', 'cancelButton');
-      await steps.verifyAbsence('ModalPage', 'overlay');
-      await steps.verifyTextContains('ModalPage', 'status', 'cancelled');
-    });
-
-    log('TC_019 Modal Page — passed');
-  });
-});
-
-test.describe('TC_020: Toast Page', () => {
-
-  test('trigger toast notifications', async ({ steps }) => {
-
-    await test.step('Navigate to Toast page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'toastLink');
-      await steps.verifyUrlContains('/toast');
-    });
-
-    await test.step('Trigger success toast and verify it appears', async () => {
-      await steps.click('ToastPage', 'successButton');
-      await steps.verifyPresence('ToastPage', 'container');
-    });
-
-    await test.step('Trigger error toast', async () => {
-      await steps.click('ToastPage', 'errorButton');
-    });
-
-    await test.step('Trigger warning toast', async () => {
-      await steps.click('ToastPage', 'warningButton');
-    });
-
-    log('TC_020 Toast Page — passed');
-  });
-});
-
-test.describe('TC_021: Tooltip & Popover Page', () => {
-
-  test('hover shows tooltip, click shows popover', async ({ steps }) => {
-
-    await test.step('Navigate to Tooltip page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tooltipLink');
-      await steps.verifyUrlContains('/tooltip');
-    });
-
-    await test.step('Hover to show tooltip', async () => {
-      await steps.hover('TooltipPage', 'tooltipTrigger1');
-      await steps.verifyPresence('TooltipPage', 'tooltipContent1');
-    });
-
-    await test.step('Click to show popover', async () => {
-      await steps.click('TooltipPage', 'popoverTrigger1');
-      await steps.verifyPresence('TooltipPage', 'popoverContent1');
-    });
-
-    log('TC_021 Tooltip & Popover Page — passed');
-  });
-});
-
-test.describe('TC_022: Drawer Page', () => {
-
-  test('open and close drawers', async ({ steps }) => {
-
-    await test.step('Navigate to Drawer page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'drawerLink');
-      await steps.verifyUrlContains('/drawer');
-    });
-
-    await test.step('Open left drawer and verify status', async () => {
-      await steps.click('DrawerPage', 'openLeftButton');
-      await steps.verifyTextContains('DrawerPage', 'status', 'left');
-    });
-
-    await test.step('Close drawer via close button', async () => {
-      await steps.click('DrawerPage', 'closeButton');
-      await steps.verifyTextContains('DrawerPage', 'status', 'closed');
-    });
-
-    await test.step('Open right drawer and close via overlay click', async () => {
-      await steps.click('DrawerPage', 'openRightButton');
-      await steps.verifyTextContains('DrawerPage', 'status', 'right');
-      await steps.click('DrawerPage', 'overlay');
-      await steps.verifyTextContains('DrawerPage', 'status', 'closed');
-    });
-
-    log('TC_022 Drawer Page — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Category 4: Widgets
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_023: Tabs Page', () => {
-
-  test('switch tabs and verify panels', async ({ steps }) => {
-
-    await test.step('Navigate to Tabs page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tabsLink');
-      await steps.verifyUrlContains('/tabs');
-    });
-
-    await test.step('Tab 1 is active by default — panel 1 visible', async () => {
-      await steps.verifyPresence('TabsPage', 'panel1');
-    });
-
-    await test.step('Click Tab 2 — panel 2 visible, panel 1 hidden', async () => {
-      await steps.click('TabsPage', 'tab2');
-      await steps.verifyPresence('TabsPage', 'panel2');
-      await steps.verifyAbsence('TabsPage', 'panel1');
-    });
-
-    await test.step('Click Tab 3 — panel 3 visible, panel 2 hidden', async () => {
-      await steps.click('TabsPage', 'tab3');
-      await steps.verifyPresence('TabsPage', 'panel3');
-      await steps.verifyAbsence('TabsPage', 'panel2');
-    });
-
-    await test.step('Click Tab 4 — panel 4 visible, panel 3 hidden', async () => {
-      await steps.click('TabsPage', 'tab4');
-      await steps.verifyPresence('TabsPage', 'panel4');
-      await steps.verifyAbsence('TabsPage', 'panel3');
-    });
-
-    await test.step('Click Tab 1 again — back to panel 1', async () => {
-      await steps.click('TabsPage', 'tab1');
-      await steps.verifyPresence('TabsPage', 'panel1');
-      await steps.verifyAbsence('TabsPage', 'panel4');
-    });
-
-    log('TC_023 Tabs Page — passed');
-  });
-});
-
-test.describe('TC_024: Accordion Page', () => {
-
-  test('expand and collapse accordion items', async ({ steps }) => {
-
-    await test.step('Navigate to Accordion page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'accordionLink');
-      await steps.verifyUrlContains('/accordion');
-    });
-
-    await test.step('Item 1 body is hidden by default', async () => {
-      await steps.verifyAbsence('AccordionPage', 'body1');
-    });
-
-    await test.step('Click header 1 — body 1 appears', async () => {
-      await steps.click('AccordionPage', 'header1');
-      await steps.verifyPresence('AccordionPage', 'body1');
-    });
-
-    await test.step('Click header 1 again — body 1 collapses', async () => {
-      await steps.click('AccordionPage', 'header1');
-      await steps.verifyAbsence('AccordionPage', 'body1');
-    });
-
-    await test.step('Expand All — all bodies visible', async () => {
-      await steps.click('AccordionPage', 'expandAllButton');
-      await steps.verifyPresence('AccordionPage', 'body1');
-      await steps.verifyPresence('AccordionPage', 'body2');
-      await steps.verifyPresence('AccordionPage', 'body3');
-    });
-
-    await test.step('Collapse All — all bodies hidden', async () => {
-      await steps.click('AccordionPage', 'collapseAllButton');
-      await steps.verifyAbsence('AccordionPage', 'body1');
-      await steps.verifyAbsence('AccordionPage', 'body2');
-      await steps.verifyAbsence('AccordionPage', 'body3');
-    });
-
-    log('TC_024 Accordion Page — passed');
-  });
-});
-
-test.describe('TC_025: Progress Page', () => {
-
-  test('verify static bars and animated progress', async ({ page, steps }) => {
-
-    await test.step('Navigate to Progress page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'progressLink');
-      await steps.verifyUrlContains('/progress');
-    });
-
-    await test.step('Verify 5 static progress bars exist', async () => {
-      await steps.verifyCount('ProgressPage', 'staticBars', { exactly: 5 });
-    });
-
-    await test.step('Animated bar starts at 0%', async () => {
-      await steps.verifyText('ProgressPage', 'animatedValue', '0%');
-    });
-
-    await test.step('Click Start — progress animates to 100%', async () => {
-      await steps.click('ProgressPage', 'startButton');
-      // Wait for animation to complete (value reaches 100%)
-      await page.locator("[data-testid='progress-animated-value']").filter({ hasText: '100%' }).waitFor({ timeout: 15000 });
-      await steps.verifyText('ProgressPage', 'animatedValue', '100%');
-    });
-
-    await test.step('Click Reset — progress resets to 0%', async () => {
-      await steps.click('ProgressPage', 'resetButton');
-      await steps.verifyText('ProgressPage', 'animatedValue', '0%');
-    });
-
-    await test.step('Circular progress value is displayed', async () => {
-      await steps.verifyPresence('ProgressPage', 'circularValue');
-    });
-
-    log('TC_025 Progress Page — passed');
-  });
-});
-
-test.describe('TC_026: Table Page', () => {
-
-  test('search, sort, paginate, and select rows', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to Table page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('Table renders 5 rows on page 1', async () => {
-      await steps.verifyCount('TablePage', 'rows', { exactly: 5 });
-    });
-
-    await test.step('Search filters rows', async () => {
-      await steps.fill('TablePage', 'searchInput', 'alice');
-      await steps.verifyCount('TablePage', 'rows', { lessThan: 5 });
-      // Clear the search
-      await steps.fill('TablePage', 'searchInput', '');
-    });
-
-    await test.step('Sort by Name column', async () => {
-      const firstNameBefore = await page.locator("[data-testid^='table-row-']:not([data-testid*='checkbox']) td:nth-child(2)").first().textContent();
-      await steps.click('TablePage', 'headerName');
-      const firstNameAfter = await page.locator("[data-testid^='table-row-']:not([data-testid*='checkbox']) td:nth-child(2)").first().textContent();
-      // After sorting, the order should change (or stay if already sorted)
-      log('Sort: %s → %s', firstNameBefore, firstNameAfter);
-    });
-
-    await test.step('Navigate to next page', async () => {
-      await steps.click('TablePage', 'nextButton');
-      await steps.verifyTextContains('TablePage', 'pageInfo', '2');
-    });
-
-    await test.step('Navigate back to previous page', async () => {
-      await steps.click('TablePage', 'prevButton');
-      await steps.verifyTextContains('TablePage', 'pageInfo', '1');
-    });
-
-    await test.step('Select a row via checkbox', async () => {
-      await steps.clickRandom('TablePage', 'rowCheckboxes');
-      await steps.verifyText('TablePage', 'selectedCount', undefined, { notEmpty: true });
-    });
-
-    await test.step('Select all rows', async () => {
-      await steps.click('TablePage', 'selectAllCheckbox');
-      await steps.verifyTextContains('TablePage', 'selectedCount', '5');
-    });
-
-    log('TC_026 Table Page — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Category 5: Interactions
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_027: Draggable Page', () => {
-
-  test('drag block and verify status updates', async ({ page, steps }) => {
-
-    await test.step('Navigate to Draggable page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'draggableLink');
-      await steps.verifyUrlContains('/draggable');
-    });
-
-    await test.step('Status starts as "none"', async () => {
-      await steps.verifyTextContains('DraggablePage', 'status', 'none');
-    });
-
-    await test.step('Drag item 1 and verify it moves', async () => {
-      await steps.dragAndDrop('DraggablePage', 'item1', { xOffset: 100, yOffset: 50 });
-    });
-
-    await test.step('All 4 draggable items are present', async () => {
-      await steps.verifyPresence('DraggablePage', 'item1');
-      await steps.verifyPresence('DraggablePage', 'item2');
-      await steps.verifyPresence('DraggablePage', 'item3');
-      await steps.verifyPresence('DraggablePage', 'item4');
-    });
-
-    log('TC_027 Draggable Page — passed');
-  });
-});
-
-test.describe('TC_028: Droppable Page', () => {
-
-  test('drop items into zones and reset', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to Droppable page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'droppableLink');
-      await steps.verifyUrlContains('/droppable');
-    });
-
-    await test.step('Initial status is Ready', async () => {
-      await steps.verifyTextContains('DroppablePage', 'status', 'Ready');
-    });
-
-    await test.step('All zones start with 0 items', async () => {
-      await steps.verifyTextContains('DroppablePage', 'redZoneCount', '0');
-      await steps.verifyTextContains('DroppablePage', 'blueZoneCount', '0');
-      await steps.verifyTextContains('DroppablePage', 'greenZoneCount', '0');
-    });
-
-    await test.step('Drag red item to red zone — correct drop', async () => {
-      const redItem = await repo.get(page, 'DroppablePage', 'redItem1');
-      const redZone = await repo.get(page, 'DroppablePage', 'redZone');
-      await steps.dragAndDrop('DroppablePage', 'redItem1', { target: redZone! });
-      await steps.verifyTextContains('DroppablePage', 'redZoneCount', '1');
-    });
-
-    await test.step('Reset returns items to source', async () => {
-      await steps.click('DroppablePage', 'resetButton');
-      await steps.verifyTextContains('DroppablePage', 'redZoneCount', '0');
-      await steps.verifyTextContains('DroppablePage', 'status', 'Ready');
-    });
-
-    log('TC_028 Droppable Page — passed');
-  });
-});
-
-test.describe('TC_029: Resizable Page', () => {
-
-  test('resize panel by dragging handle', async ({ steps }) => {
-
-    await test.step('Navigate to Resizable page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'resizableLink');
-      await steps.verifyUrlContains('/resizable');
-    });
-
-    await test.step('Initial width is 300px', async () => {
-      await steps.verifyTextContains('ResizablePage', 'widthDisplay', '300');
-    });
-
-    await test.step('Drag handle right to increase width', async () => {
-      await steps.dragAndDrop('ResizablePage', 'handle', { xOffset: 100, yOffset: 0 });
-    });
-
-    await test.step('Width display is present and updated', async () => {
-      await steps.verifyText('ResizablePage', 'widthDisplay', undefined, { notEmpty: true });
-    });
-
-    log('TC_029 Resizable Page — passed');
-  });
-});
-
-test.describe('TC_030: Kanban Page', () => {
-
-  test('add card and verify columns', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to Kanban page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'kanbanLink');
-      await steps.verifyUrlContains('/kanban');
-    });
-
-    await test.step('All three columns are present', async () => {
-      await steps.verifyPresence('KanbanPage', 'todoColumn');
-      await steps.verifyPresence('KanbanPage', 'inProgressColumn');
-      await steps.verifyPresence('KanbanPage', 'doneColumn');
-    });
-
-    await test.step('Cards are present initially', async () => {
-      await steps.verifyCount('KanbanPage', 'cards', { greaterThan: 0 });
-    });
-
-    await test.step('Add a new card to Todo column', async () => {
-      const allCards = await repo.getAll(page, 'KanbanPage', 'cards');
-      const countBefore = allCards!.length;
-      await steps.click('KanbanPage', 'addTodoButton');
-      await steps.verifyCount('KanbanPage', 'cards', { greaterThan: countBefore });
-    });
-
-    log('TC_030 Kanban Page — passed');
-  });
-});
-
-test.describe('TC_031: Infinite Scroll Page', () => {
-
-  test('scroll to load more items', async ({ page, steps }) => {
-
-    await test.step('Navigate to Infinite Scroll page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'infiniteScrollLink');
-      await steps.verifyUrlContains('/infinite-scroll');
-    });
-
-    await test.step('Wait for initial items to load', async () => {
-      await page.locator("[data-testid^='scroll-item-']").first().waitFor({ timeout: 10000 });
-      await steps.verifyCount('InfiniteScrollPage', 'items', { greaterThan: 0 });
-    });
-
-    await test.step('Scroll container to trigger more loads', async () => {
-      const container = page.locator("[data-testid='scroll-container']");
-      // Scroll to the bottom of the container
-      await container.evaluate((el) => { el.scrollTop = el.scrollHeight; });
-      // Wait for more items to load
-      await page.waitForTimeout(2000);
-      const countAfterScroll = await page.locator("[data-testid^='scroll-item-']").count();
-      expect(countAfterScroll).toBeGreaterThan(10);
-    });
-
-    log('TC_031 Infinite Scroll Page — passed');
-  });
-});
-
-test.describe('TC_032: Loading States Page', () => {
-
-  test('verify spinner, skeleton toggle, and button loading', async ({ steps }) => {
-
-    await test.step('Navigate to Loading States page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'loadingLink');
-      await steps.verifyUrlContains('/loading');
-    });
-
-    await test.step('Spinner is visible', async () => {
-      await steps.verifyPresence('LoadingStatesPage', 'spinner');
-    });
-
-    await test.step('Skeleton is visible by default', async () => {
-      await steps.verifyPresence('LoadingStatesPage', 'skeleton');
-    });
-
-    await test.step('Toggle skeleton off', async () => {
-      await steps.click('LoadingStatesPage', 'skeletonToggle');
-      await steps.verifyAbsence('LoadingStatesPage', 'skeleton');
-    });
-
-    await test.step('Toggle skeleton back on', async () => {
-      await steps.click('LoadingStatesPage', 'skeletonToggle');
-      await steps.verifyPresence('LoadingStatesPage', 'skeleton');
-    });
-
-    await test.step('Click loading button — enters loading state', async () => {
-      await steps.click('LoadingStatesPage', 'loadingButton');
-      // Button should show loading state (text changes or spinner appears)
-      await steps.verifyPresence('LoadingStatesPage', 'loadingButton');
-    });
-
-    log('TC_032 Loading States Page — passed');
-  });
-});
-
-test.describe('TC_033: Dynamic Form Page', () => {
-
-  test('add fields and submit form', async ({ page, steps }) => {
-
-    await test.step('Navigate to Dynamic Form page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'dynamicFormLink');
-      await steps.verifyUrlContains('/dynamic-form');
-    });
-
-    await test.step('Starts with 1 field', async () => {
-      await steps.verifyCount('DynamicFormPage', 'fields', { exactly: 1 });
-    });
-
-    await test.step('Add a second field', async () => {
-      await steps.click('DynamicFormPage', 'addButton');
-      await steps.verifyCount('DynamicFormPage', 'fields', { exactly: 2 });
-    });
-
-    await test.step('Add a third field', async () => {
-      await steps.click('DynamicFormPage', 'addButton');
-      await steps.verifyCount('DynamicFormPage', 'fields', { exactly: 3 });
-    });
-
-    await test.step('Fill field 1 and submit', async () => {
-      await steps.fill('DynamicFormPage', 'field1', 'Test Value');
-      await steps.click('DynamicFormPage', 'submitButton');
-    });
-
-    log('TC_033 Dynamic Form Page — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Category 6: Media
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_034: Gallery Page', () => {
-
-  test('verify gallery images and lightbox', async ({ page, steps }) => {
-
-    await test.step('Navigate to Gallery page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'galleryLink');
-      await steps.verifyUrlContains('/gallery');
-    });
-
-    await test.step('8 gallery items rendered', async () => {
-      const items = page.locator("[data-testid^='gallery-item-']");
-      await expect(items).toHaveCount(8);
-    });
-
-    await test.step('Click image to open lightbox', async () => {
-      await page.locator("[data-testid='gallery-item-1']").click();
-      const overlay = page.locator("[data-testid='gallery-overlay']");
-      await overlay.waitFor({ state: 'visible', timeout: 5000 });
-    });
-
-    await test.step('Close lightbox', async () => {
-      const closeButton = page.locator("[data-testid='gallery-close']");
-      await closeButton.click();
-      const overlay = page.locator("[data-testid='gallery-overlay']");
-      await overlay.waitFor({ state: 'hidden', timeout: 5000 });
-    });
-
-    log('TC_034 Gallery Page — passed');
-  });
-});
-
-test.describe('TC_035: Carousel Page', () => {
-
-  test('navigate slides with buttons and dots', async ({ steps }) => {
-
-    await test.step('Navigate to Carousel page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'carouselLink');
-      await steps.verifyUrlContains('/carousel');
-    });
-
-    await test.step('First slide visible on load', async () => {
-      await steps.verifyPresence('CarouselPage', 'slide1');
-    });
-
-    await test.step('Click Next — slide advances', async () => {
-      await steps.click('CarouselPage', 'nextButton');
-      await steps.verifyPresence('CarouselPage', 'slide2');
-    });
-
-    await test.step('Click Previous — goes back', async () => {
-      await steps.click('CarouselPage', 'prevButton');
-      await steps.verifyPresence('CarouselPage', 'slide1');
-    });
-
-    await test.step('Click dot 2 — jumps to slide 2', async () => {
-      await steps.click('CarouselPage', 'dot2');
-      await steps.verifyPresence('CarouselPage', 'slide2');
-    });
-
-    await test.step('Autoplay toggle is present', async () => {
-      await steps.verifyPresence('CarouselPage', 'autoplayButton');
-    });
-
-    log('TC_035 Carousel Page — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Category 7: Auth & State
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_036: Login Form Page', () => {
-
-  test('validation errors and successful login', async ({ page, steps }) => {
-
-    await test.step('Navigate to Login Form page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'loginFormLink');
-      await steps.verifyUrlContains('/login-form');
-    });
-
-    await test.step('Empty submit shows validation errors', async () => {
-      await steps.click('LoginFormPage', 'signInButton');
-      // Validation errors should appear for empty fields
-      const errors = page.locator("[data-testid='login-username-error'], [data-testid='login-password-error']");
-      await errors.first().waitFor({ state: 'visible', timeout: 5000 });
-    });
-
-    await test.step('Wrong credentials show error', async () => {
-      await steps.fill('LoginFormPage', 'usernameInput', 'wrong');
-      await steps.fill('LoginFormPage', 'passwordInput', 'wrongpass');
-      await steps.click('LoginFormPage', 'signInButton');
-      const loginError = page.locator("[data-testid='login-error']");
-      await loginError.waitFor({ state: 'visible', timeout: 5000 });
-    });
-
-    await test.step('Show/hide password toggle', async () => {
-      const passwordField = page.locator('#login-password-input');
-      expect(await passwordField.getAttribute('type')).toBe('password');
-      await steps.click('LoginFormPage', 'showPasswordButton');
-      expect(await passwordField.getAttribute('type')).toBe('text');
-      await steps.click('LoginFormPage', 'showPasswordButton');
-      expect(await passwordField.getAttribute('type')).toBe('password');
-    });
-
-    await test.step('Correct login succeeds', async () => {
-      await steps.fill('LoginFormPage', 'usernameInput', 'admin');
-      await steps.fill('LoginFormPage', 'passwordInput', 'password123');
-      await steps.click('LoginFormPage', 'signInButton');
-      const success = page.locator("[data-testid='login-success']");
-      await success.waitFor({ state: 'visible', timeout: 5000 });
-    });
-
-    log('TC_036 Login Form Page — passed');
-  });
-});
-
-test.describe('TC_037: Pinia Counter Page', () => {
-
-  test('increment, decrement, reset, and change step', async ({ steps }) => {
-
-    await test.step('Navigate to Pinia Counter page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'piniaCounterLink');
-      await steps.verifyUrlContains('/pinia-counter');
-    });
-
-    await test.step('Initial value is 0', async () => {
-      await steps.verifyText('PiniaCounterPage', 'counterValue', '0');
-    });
-
-    await test.step('Increment increases value', async () => {
-      await steps.click('PiniaCounterPage', 'incrementButton');
-      await steps.verifyText('PiniaCounterPage', 'counterValue', '1');
-    });
-
-    await test.step('Decrement decreases value', async () => {
-      await steps.click('PiniaCounterPage', 'decrementButton');
-      await steps.verifyText('PiniaCounterPage', 'counterValue', '0');
-    });
-
-    await test.step('Increment twice and reset', async () => {
-      await steps.click('PiniaCounterPage', 'incrementButton');
-      await steps.click('PiniaCounterPage', 'incrementButton');
-      await steps.verifyText('PiniaCounterPage', 'counterValue', '2');
-      await steps.click('PiniaCounterPage', 'resetButton');
-      await steps.verifyText('PiniaCounterPage', 'counterValue', '0');
-    });
-
-    await test.step('Change step to 5 and increment', async () => {
-      await steps.fill('PiniaCounterPage', 'stepInput', '5');
-      await steps.click('PiniaCounterPage', 'incrementButton');
-      await steps.verifyText('PiniaCounterPage', 'counterValue', '5');
-    });
-
-    await test.step('History tracks operations', async () => {
-      await steps.verifyPresence('PiniaCounterPage', 'history');
-      await steps.verifyText('PiniaCounterPage', 'history', undefined, { notEmpty: true });
-    });
-
-    log('TC_037 Pinia Counter Page — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Category 8: Edge Cases
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_038: Long List Page', () => {
-
-  test('renders 200 items with search filtering', async ({ page, steps }) => {
-
-    await test.step('Navigate to Long List page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'longListLink');
-      await steps.verifyUrlContains('/long-list');
-    });
-
-    await test.step('200 items are rendered', async () => {
-      await steps.verifyCount('LongListPage', 'listItems', { exactly: 200 });
-    });
-
-    await test.step('Search filters the list', async () => {
-      await steps.fill('LongListPage', 'searchInput', 'Item 1');
-      await steps.verifyCount('LongListPage', 'listItems', { lessThan: 200 });
-    });
-
-    await test.step('Clear search restores full list', async () => {
-      await steps.fill('LongListPage', 'searchInput', '');
-      await steps.verifyCount('LongListPage', 'listItems', { exactly: 200 });
-    });
-
-    log('TC_038 Long List Page — passed');
-  });
-});
-
-test.describe('TC_039: Multi-step Form Page', () => {
-
-  test('navigate through form steps', async ({ page, steps }) => {
-
-    await test.step('Navigate to Multi-step Form page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'multistepLink');
-      await steps.verifyUrlContains('/multistep');
-    });
-
-    await test.step('Starts at step 1', async () => {
-      await steps.verifyTextContains('MultiStepFormPage', 'currentStep', '1');
-    });
-
-    await test.step('Fill step 1 and go to step 2', async () => {
-      await steps.fill('MultiStepFormPage', 'firstNameInput', 'John');
-      await steps.fill('MultiStepFormPage', 'lastNameInput', 'Doe');
-      await steps.click('MultiStepFormPage', 'nextButton');
-      await steps.verifyTextContains('MultiStepFormPage', 'currentStep', '2');
-    });
-
-    await test.step('Fill step 2 and go to step 3', async () => {
-      await steps.fill('MultiStepFormPage', 'emailInput', 'john@example.com');
-      await steps.fill('MultiStepFormPage', 'phoneInput', '555-1234');
-      await steps.click('MultiStepFormPage', 'nextButton');
-      await steps.verifyTextContains('MultiStepFormPage', 'currentStep', '3');
-    });
-
-    await test.step('Step 3 has message input', async () => {
-      await steps.verifyPresence('MultiStepFormPage', 'messageInput');
-    });
-
-    await test.step('Go back to step 2 and return', async () => {
-      await steps.click('MultiStepFormPage', 'prevButton');
-      await steps.verifyTextContains('MultiStepFormPage', 'currentStep', '2');
-      await steps.click('MultiStepFormPage', 'nextButton');
-      await steps.verifyTextContains('MultiStepFormPage', 'currentStep', '3');
-    });
-
-    await test.step('Fill message and submit — summary appears', async () => {
-      await steps.fill('MultiStepFormPage', 'messageInput', 'Test message');
-      await steps.click('MultiStepFormPage', 'submitButton');
-      const submitted = page.locator('h2:has-text("Submitted!")');
-      await submitted.waitFor({ state: 'visible', timeout: 5000 });
-    });
-
-    log('TC_039 Multi-step Form Page — passed');
-  });
-});
-
-test.describe('TC_040: State Viewer Page', () => {
-
-  test('switch between states', async ({ steps }) => {
-
-    await test.step('Navigate to State Viewer page via sidebar', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'stateViewerLink');
-      await steps.verifyUrlContains('/state-viewer');
-    });
-
-    await test.step('Click Empty state button', async () => {
-      await steps.click('StateViewerPage', 'emptyButton');
-      await steps.verifyPresence('StateViewerPage', 'emptyView');
-      await steps.verifyTextContains('StateViewerPage', 'currentState', 'empty');
-    });
-
-    await test.step('Click Loading state button', async () => {
-      await steps.click('StateViewerPage', 'loadingButton');
-      await steps.verifyPresence('StateViewerPage', 'loadingView');
-      await steps.verifyTextContains('StateViewerPage', 'currentState', 'loading');
-    });
-
-    await test.step('Click Error state button', async () => {
-      await steps.click('StateViewerPage', 'errorButton');
-      await steps.verifyPresence('StateViewerPage', 'errorView');
-      await steps.verifyTextContains('StateViewerPage', 'currentState', 'error');
-    });
-
-    await test.step('Click Populated state button', async () => {
-      await steps.click('StateViewerPage', 'populatedButton');
-      await steps.verifyTextContains('StateViewerPage', 'currentState', 'populated');
-    });
-
-    log('TC_040 State Viewer Page — passed');
-  });
-});
-
-// ══════════════════════════════════════════════════════════════════════════════
-// 100% API Coverage Tests
-// ══════════════════════════════════════════════════════════════════════════════
-
-test.describe('TC_041: Steps - Navigation, Viewport & Scroll', () => {
-
-  test('backOrForward, refresh, setViewport, scrollIntoView, pressKey', async ({ page, steps }) => {
-
-    await test.step('Navigate to Buttons page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-    });
-
-    await test.step('refresh reloads the page', async () => {
-      await steps.refresh();
-      await steps.verifyPresence('ButtonsPage', 'primaryButton');
-    });
-
-    await test.step('Navigate to a second page then backOrForward', async () => {
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-      await steps.backOrForward('BACKWARDS');
-      await steps.verifyUrlContains('/buttons');
-      await steps.backOrForward('FORWARDS');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('scrollIntoView scrolls an element into view', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-      await steps.scrollIntoView('ButtonsPage', 'loadingButton');
-    });
-
-    await test.step('pressKey sends a keyboard event', async () => {
-      await steps.pressKey('Tab');
-    });
-
-    await test.step('setViewport changes the viewport size', async () => {
-      await steps.setViewport(800, 600);
-      const size = page.viewportSize();
-      expect(size?.width).toBe(800);
-      expect(size?.height).toBe(600);
-    });
-
-    log('TC_041 Steps Navigation & Viewport — passed');
-  });
-});
-
-test.describe('TC_042: Steps - Click Variants & Data Extraction', () => {
-
-  test('clickIfPresent, clickWithoutScrolling, getText, getAttribute, verifyAttribute', async ({ steps }) => {
-
-    await test.step('Navigate to Buttons page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-    });
-
-    await test.step('clickWithoutScrolling clicks without scrolling', async () => {
-      await steps.clickWithoutScrolling('ButtonsPage', 'primaryButton');
-      await steps.verifyTextContains('ButtonsPage', 'resultText', 'Primary');
-    });
-
-    await test.step('clickIfPresent clicks a present element', async () => {
-      await steps.clickIfPresent('ButtonsPage', 'secondaryButton');
-      await steps.verifyTextContains('ButtonsPage', 'resultText', 'Secondary');
-    });
-
-    await test.step('clickIfPresent does nothing for a missing element', async () => {
-      await steps.clickIfPresent('FormsPage', 'title'); // not on this page
-    });
-
-    await test.step('getText returns element text content', async () => {
-      const text = await steps.getText('ButtonsPage', 'resultText');
-      expect(text).toContain('Secondary');
-    });
-
-    await test.step('getAttribute returns an attribute value', async () => {
-      const testId = await steps.getAttribute('ButtonsPage', 'disabledButton', 'data-testid');
-      expect(testId).toBe('btn-disabled');
-    });
-
-    await test.step('verifyAttribute asserts an attribute value', async () => {
-      await steps.verifyAttribute('ButtonsPage', 'disabledButton', 'data-testid', 'btn-disabled');
-    });
-
-    log('TC_042 Click Variants & Data Extraction — passed');
-  });
-});
-
-test.describe('TC_043: Steps - Tab Management', () => {
-
-  test('closeTab and getTabCount', async ({ page, steps }) => {
-
-    await test.step('Navigate to Alerts page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'alertsLink');
-      await steps.verifyUrlContains('/alerts');
-    });
-
-    await test.step('getTabCount returns 1 initially', async () => {
-      const count = steps.getTabCount();
-      expect(count).toBe(1);
-    });
-
-    await test.step('Open new tab and close it via steps.closeTab', async () => {
-      const newPage = await steps.switchToNewTab(async () => {
-        await steps.click('AlertsPage', 'newTabButton');
-      });
-      expect(steps.getTabCount()).toBe(2);
-
-      await steps.closeTab(newPage);
-      expect(steps.getTabCount()).toBe(1);
-    });
-
-    log('TC_043 Tab Management — passed');
-  });
-});
-
-test.describe('TC_044: Repo - getRandom & setDefaultTimeout', () => {
-
-  test('getRandom picks a random element, setDefaultTimeout sets timeout', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('setDefaultTimeout changes the repo default timeout', async () => {
-      repo.setDefaultTimeout(10000);
-      // Just verify it does not throw
-    });
-
-    await test.step('getRandom returns one of the category cards', async () => {
-      const randomCard = await repo.getRandom(page, 'HomePage', 'categories');
-      expect(randomCard).toBeTruthy();
-      const text = await randomCard!.textContent();
-      expect(text).toBeTruthy();
-    });
-
-    log('TC_044 Repo getRandom & setDefaultTimeout — passed');
-  });
-});
-
-test.describe('TC_045: ContextStore - Full API', () => {
-
-  test('has, remove, clear, getBoolean, getNumber, items, merge', async ({ contextStore }) => {
-
-    await test.step('has returns false for missing key', () => {
-      expect(contextStore.has('nonexistent')).toBe(false);
-    });
-
-    await test.step('put + has returns true', () => {
-      contextStore.put('key1', 'value1');
-      expect(contextStore.has('key1')).toBe(true);
-    });
-
-    await test.step('remove deletes a key', () => {
-      contextStore.remove('key1');
-      expect(contextStore.has('key1')).toBe(false);
-    });
-
-    await test.step('getBoolean returns correct boolean', () => {
-      contextStore.put('flag', 'true');
-      expect(contextStore.getBoolean('flag')).toBe(true);
-      contextStore.put('flag2', 'false');
-      expect(contextStore.getBoolean('flag2')).toBe(false);
-      expect(contextStore.getBoolean('missing', false)).toBe(false);
-    });
-
-    await test.step('getNumber returns correct number', () => {
-      contextStore.put('count', '42');
-      expect(contextStore.getNumber('count')).toBe(42);
-      expect(contextStore.getNumber('missing', 0)).toBe(0);
-    });
-
-    await test.step('items returns a Set of keys', () => {
-      contextStore.clear();
-      contextStore.put('a', '1');
-      contextStore.put('b', '2');
-      const keys = contextStore.items();
-      expect(keys.has('a')).toBe(true);
-      expect(keys.has('b')).toBe(true);
-      expect(keys.size).toBe(2);
-    });
-
-    await test.step('merge merges a record into the store', () => {
-      contextStore.clear();
-      contextStore.merge({ x: '10', y: '20' });
-      expect(contextStore.get('x')).toBe('10');
-      expect(contextStore.get('y')).toBe('20');
-    });
-
-    await test.step('clear removes all entries', () => {
-      contextStore.clear();
-      expect(contextStore.items().size).toBe(0);
-    });
-
-    log('TC_045 ContextStore Full API — passed');
-  });
-});
-
-test.describe('TC_046: verifyImages - Image Verification', () => {
-
-  test('verifyImages on product carousel and thumbnail grid', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to Product Carousel page', async () => {
-      await steps.navigateTo('/'); 
-      await steps.click('SidebarNav', 'productCarouselLink');
-      await steps.verifyUrlContains('/product-carousel');
-    });
-
-    await test.step('verifyImages on the active carousel slide', async () => {
-      await steps.verifyImages('ProductCarouselPage', 'firstCarouselImage');
-    });
-
-    await test.step('verifyImages on the product grid thumbnails', async () => {
-      await steps.verifyImages('ProductCarouselPage', 'productThumbnails');
-    });
-
-    log('TC_046 verifyImages — passed');
-  });
-});
-
-// ══════════════════════════════════════════════════════════════════════════════
-// Advanced/Raw ElementInteractions Sub-API Tests
-// ══════════════════════════════════════════════════════════════════════════════
-
-test.describe('TC_047: Raw Interactions - interact methods', () => {
-
-  test('all interact sub-API methods', async ({ page, repo, steps, interactions }) => {
-
-    await test.step('Navigate to Buttons page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-    });
-
-    await test.step('interact.click', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'primaryButton');
-      await interactions.interact.click(btn!);
-    });
-
-    await test.step('interact.clickWithoutScrolling', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'secondaryButton');
-      await interactions.interact.clickWithoutScrolling(btn!);
-    });
-
-    await test.step('interact.clickIfPresent', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'dangerButton');
-      await interactions.interact.clickIfPresent(btn!);
-    });
-
-    await test.step('interact.doubleClick', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'primaryButton');
-      await interactions.interact.doubleClick(btn!);
-    });
-
-    await test.step('interact.rightClick', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'primaryButton');
-      await interactions.interact.rightClick(btn!);
-    });
-
-    await test.step('interact.hover', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'primaryButton');
-      await interactions.interact.hover(btn!);
-    });
-
-    await test.step('interact.pressKey', async () => {
-      await interactions.interact.pressKey('Escape');
-    });
-
-    await test.step('Navigate to Text Inputs for fill/type tests', async () => {
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('interact.fill', async () => {
-      const input = await repo.get(page, 'TextInputsPage', 'textInput');
-      await interactions.interact.fill(input!, 'raw fill test');
-    });
-
-    await test.step('interact.typeSequentially', async () => {
-      const input = await repo.get(page, 'TextInputsPage', 'textInput');
-      await interactions.interact.fill(input!, '');
-      await interactions.interact.typeSequentially(input!, 'raw typed');
-    });
-
-    await test.step('Navigate to Checkboxes for check/uncheck', async () => {
-      await steps.click('SidebarNav', 'checkboxesLink');
-      await steps.verifyUrlContains('/checkboxes');
-    });
-
-    await test.step('interact.check', async () => {
-      const cb = await repo.get(page, 'CheckboxesPage', 'uncheckedCheckbox');
-      await interactions.interact.check(cb!);
-    });
-
-    await test.step('interact.uncheck', async () => {
-      const cb = await repo.get(page, 'CheckboxesPage', 'uncheckedCheckbox');
-      await interactions.interact.uncheck(cb!);
-    });
-
-    await test.step('Navigate to Sliders for setSliderValue', async () => {
-      await steps.click('SidebarNav', 'slidersLink');
-      await steps.verifyUrlContains('/sliders');
-    });
-
-    await test.step('interact.setSliderValue', async () => {
-      const slider = await repo.get(page, 'SlidersPage', 'basicSlider');
-      await interactions.interact.setSliderValue(slider!, 60);
-    });
-
-    await test.step('interact.scrollIntoView', async () => {
-      const slider = await repo.get(page, 'SlidersPage', 'disabledSlider');
-      await interactions.interact.scrollIntoView(slider!);
-    });
-
-    await test.step('Navigate to Dropdown for selectDropdown', async () => {
-      await steps.click('SidebarNav', 'dropdownLink');
-      await steps.verifyUrlContains('/dropdown');
-    });
-
-    await test.step('interact.selectDropdown', async () => {
-      const select = await repo.get(page, 'DropdownSelectPage', 'singleSelect');
-      const value = await interactions.interact.selectDropdown(select!);
-      expect(value).toBeTruthy();
-    });
-
-    await test.step('Navigate to File Upload for uploadFile', async () => {
-      await steps.click('SidebarNav', 'fileUploadLink');
-      await steps.verifyUrlContains('/file-upload');
-    });
-
-    await test.step('interact.uploadFile', async () => {
-      const input = await repo.get(page, 'FileUploadPage', 'singleFileInput');
-      await interactions.interact.uploadFile(input!, 'tests/fixtures/test-upload.txt');
-    });
-
-    await test.step('Navigate to Draggable for dragAndDrop', async () => {
-      await steps.click('SidebarNav', 'draggableLink');
-      await steps.verifyUrlContains('/draggable');
-    });
-
-    await test.step('interact.dragAndDrop', async () => {
-      const item = await repo.get(page, 'DraggablePage', 'item1');
-      await interactions.interact.dragAndDrop(item!, { xOffset: 50, yOffset: 0 });
-    });
-
-    // --- NEW OTP TESTS START HERE ---
-
-    await test.step('Navigate to OTP page for character-by-character testing', async () => {
-      await steps.click('SidebarNav', 'otpLink');
-      await steps.verifyUrlContains('/otp');
-    });
-
-    await test.step('interact.click - Increment OTP length', async () => {
-      // Testing the '+' button on the Basic section
-      const addDigitBtn = await repo.get(page, 'OtpPage', 'addBasicDigitBtn');
-      await interactions.interact.click(addDigitBtn!);
-    });
-
-    await test.step('interact.getText - Read dynamically generated code', async () => {
-      // Testing data extraction from the generated code block
-      const generatedCodeBlock = await repo.get(page, 'OtpPage', 'basicGeneratedCode');
-      const code = await interactions.extract.getText(generatedCodeBlock!);
-      expect(code).toBeTruthy();
-      expect(code!.length).toBeGreaterThan(0);
-    });
-
-    await test.step('interact.typeSequentially - Enter OTP code naturally', async () => {
-      // Utilizing typeSequentially for its ideal use case: OTP inputs
-      const otpInput = await repo.get(page, 'OtpPage', 'basicOtpInput');
-      // Types '12345' with a 50ms delay between keystrokes to mimic human entry
-      await interactions.interact.typeSequentially(otpInput!, '12345', 50);
-    });
-
-    // --- NEW OTP TESTS END HERE ---
-
-    await test.step('interact.getByText', async () => {
-      await steps.navigateTo('/');
-      const card = await repo.get(page, 'HomePage', 'categories');
-      const result = await interactions.interact.getByText(card!, 'HomePage', 'categories', 'Elements');
-      expect(result).toBeTruthy();
-    });
-
-    log('TC_047 Raw interact methods — passed');
-  });
-});
-
-test.describe('TC_048: Raw Interactions - verify methods', () => {
-
-  test('all verify sub-API methods', async ({ page, repo, steps, interactions }) => {
-
-    await test.step('Navigate to Buttons page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-    });
-
-    await test.step('verify.presence', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'primaryButton');
-      await interactions.verify.presence(btn!);
-    });
-
-    await test.step('verify.absence', async () => {
-      const formsTitle = page.locator("[data-testid='forms-title-absent']");
-      await interactions.verify.absence(formsTitle);
-    });
-
-    await test.step('verify.text (exact)', async () => {
-      await steps.click('ButtonsPage', 'primaryButton');
-      const result = await repo.get(page, 'ButtonsPage', 'resultText');
-      await interactions.verify.text(result!, 'Primary');
-    });
-
-    await test.step('verify.textContains', async () => {
-      const result = await repo.get(page, 'ButtonsPage', 'resultText');
-      await interactions.verify.textContains(result!, 'Primary');
-    });
-
-    await test.step('verify.state', async () => {
-      const disabled = await repo.get(page, 'ButtonsPage', 'disabledButton');
-      await interactions.verify.state(disabled!, 'disabled');
-    });
-
-    await test.step('verify.attribute', async () => {
-      const disabled = await repo.get(page, 'ButtonsPage', 'disabledButton');
-      await interactions.verify.attribute(disabled!, 'data-testid', 'btn-disabled');
-    });
-
-    await test.step('verify.count', async () => {
-      await steps.navigateTo('/');
-      const categoriesLocator = page.locator(repo.getSelector('HomePage', 'categories'));
-      await interactions.verify.count(categoriesLocator, { exactly: 8 });
-    });
-
-    await test.step('verify.urlContains', async () => {
-      await steps.navigateTo('/otp');
-      await interactions.verify.urlContains('/otp');
-    });
-
-    await test.step('verify.tabCount', async () => {
-      await interactions.verify.tabCount(1);
-    });
-
-    await test.step('Navigate to Text Inputs for inputValue', async () => {
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('verify.inputValue', async () => {
-      const input = await repo.get(page, 'TextInputsPage', 'textInput');
-      await interactions.interact.fill(input!, 'test value');
-      await interactions.verify.inputValue(input!, 'test value');
-    });
-
-    await test.step('verify.images (negative — no img elements)', async () => {
-      await steps.click('SidebarNav', 'galleryLink');
-      await steps.verifyUrlContains('/gallery');
-      const items = page.locator("[data-testid^='gallery-item-'] img");
-      let errorCaught = false;
-      try {
-        await interactions.verify.images(items, false);
-      } catch {
-        errorCaught = true;
-      }
-      log('verify.images exercised, errorCaught=%s', errorCaught);
-    });
-
-    log('TC_048 Raw verify methods — passed');
-  });
-});
-
-test.describe('TC_049: Raw Interactions - extract & navigate methods', () => {
-
-  test('all extract and navigate sub-API methods', async ({ page, repo, steps, interactions }) => {
-
-    await test.step('Navigate to Buttons page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-    });
-
-    await test.step('extract.getText', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'primaryButton');
-      const text = await interactions.extract.getText(btn!);
-      expect(text).toContain('Primary');
-    });
-
-    await test.step('extract.getAttribute', async () => {
-      const btn = await repo.get(page, 'ButtonsPage', 'disabledButton');
-      const attr = await interactions.extract.getAttribute(btn!, 'data-testid');
-      expect(attr).toBe('btn-disabled');
-    });
-
-    await test.step('navigate.toUrl', async () => {
-      await interactions.navigate.toUrl('/');
-      await steps.verifyCount('HomePage', 'categories', { exactly: 8 });
-    });
-
-    await test.step('navigate.setViewport', async () => {
-      await interactions.navigate.setViewport(1024, 768);
-      const size = page.viewportSize();
-      expect(size?.width).toBe(1024);
-    });
-
-    await test.step('navigate.reload', async () => {
-      await interactions.navigate.reload();
-      await steps.verifyCount('HomePage', 'categories', { exactly: 8 });
-    });
-
-    await test.step('navigate.backOrForward', async () => {
-      await steps.click('SidebarNav', 'buttonsLink');
-      await steps.verifyUrlContains('/buttons');
-      await interactions.navigate.backOrForward('BACKWARDS');
-      await steps.verifyCount('HomePage', 'categories', { exactly: 8 });
-    });
-
-    await test.step('navigate.getTabCount', async () => {
-      const count = interactions.navigate.getTabCount();
-      expect(count).toBe(1);
-    });
-
-    await test.step('navigate.switchToNewTab + navigate.closeTab', async () => {
-      await steps.click('SidebarNav', 'alertsLink');
-      await steps.verifyUrlContains('/alerts');
-
-      const newPage = await interactions.navigate.switchToNewTab(async () => {
-        await steps.click('AlertsPage', 'newTabButton');
-      });
-      expect(interactions.navigate.getTabCount()).toBe(2);
-
-      await interactions.navigate.closeTab(newPage);
-      expect(interactions.navigate.getTabCount()).toBe(1);
-    });
-
-    log('TC_049 Raw extract & navigate methods — passed');
-  });
-});
-
-test.describe('TC_050: Repo - getByAttribute, getByIndex, getByRole, getVisible', () => {
-
-  test('new repo accessor methods', async ({ page, repo, steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('getByIndex returns the element at a given index', async () => {
-      const secondCard = await repo.getByIndex(page, 'HomePage', 'categories', 1);
-      expect(secondCard).toBeTruthy();
-      const text = await secondCard!.textContent();
-      expect(text).toBeTruthy();
-    });
-
-    await test.step('getByIndex returns null for out-of-bounds index', async () => {
-      const missing = await repo.getByIndex(page, 'HomePage', 'categories', 999);
-      expect(missing).toBeNull();
-    });
-
-    await test.step('getVisible returns first visible element', async () => {
-      const visible = await repo.getVisible(page, 'HomePage', 'categories');
-      expect(visible).toBeTruthy();
-      const text = await visible!.textContent();
-      expect(text).toBeTruthy();
-    });
-
-    await test.step('getByAttribute finds element by data-testid', async () => {
-      const formsCard = await repo.getByAttribute(page, 'HomePage', 'categories', 'data-testid', 'home-card-forms');
-      expect(formsCard).toBeTruthy();
-      const text = await formsCard!.textContent();
-      expect(text).toContain('Forms');
-    });
-
-    await test.step('getByRole finds element by explicit role attribute', async () => {
-      // getByRole checks the explicit HTML `role` attribute
-      // Category cards may not have explicit role — result may be null
-      const result = await repo.getByRole(page, 'HomePage', 'categories', 'link');
-      // Just exercise the method — whether found or null, coverage is achieved
-      log('getByRole result: %s', result ? 'found' : 'null');
-    });
-
-    log('TC_050 Repo accessor methods — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// Listed Element Methods
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_051: clickListedElement', () => {
-
-  test('all identification and child-targeting variants', async ({ steps }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('Click by text', async () => {
-      await steps.clickListedElement('TablePage', 'rows', { text: 'Alice' });
-    });
-
-    await test.step('Click by text + child CSS selector', async () => {
-      await steps.clickListedElement('TablePage', 'rows', {
-        text: 'Alice',
-        child: 'td:nth-child(2)'
-      });
-    });
-
-    await test.step('Click by attribute', async () => {
-      await steps.navigateTo('/');
-      await steps.clickListedElement('HomePage', 'categories', {
-        attribute: { name: 'data-testid', value: 'home-card-forms' }
-      });
-      await steps.verifyUrlContains('/forms');
-    });
-
-    await test.step('Click by attribute + child CSS selector', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-      await steps.clickListedElement('TablePage', 'rows', {
-        attribute: { name: 'data-testid', value: 'table-row-1' },
-        child: 'td:nth-child(2)'
-      });
-    });
-
-    log('TC_051 clickListedElement — passed');
-  });
-});
-
-test.describe('TC_052: verifyListedElement', () => {
-
-  test('all verification variants', async ({ steps }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('Verify presence by text (no assertion fields)', async () => {
-      await steps.verifyListedElement('TablePage', 'rows', { text: 'Alice' });
-    });
-
-    await test.step('Verify expectedText on matched element (no child)', async () => {
-      await steps.navigateTo('/');
-      await steps.verifyListedElement('HomePage', 'categories', {
-        text: 'Forms',
-        expectedText: 'Forms5 components'
-      });
-    });
-
-    await test.step('Verify expectedText on a child element (text + child + expectedText)', async () => {
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-      await steps.verifyListedElement('TablePage', 'rows', {
-        text: 'Alice',
-        child: 'td:nth-child(2)',
-        expectedText: 'Alice Martin'
-      });
-    });
-
-    await test.step('Verify expected attribute by text match', async () => {
-      await steps.verifyListedElement('TablePage', 'rows', {
-        text: 'Alice',
-        expected: { name: 'data-testid', value: 'table-row-1' }
-      });
-    });
-
-    await test.step('Verify expected attribute by attribute match', async () => {
-      await steps.navigateTo('/');
-      await steps.verifyListedElement('HomePage', 'categories', {
-        attribute: { name: 'data-testid', value: 'home-card-forms' },
-        expected: { name: 'data-testid', value: 'home-card-forms' }
-      });
-    });
-
-    await test.step('Verify expectedText on child via attribute match', async () => {
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-      await steps.verifyListedElement('TablePage', 'rows', {
-        attribute: { name: 'data-testid', value: 'table-row-1' },
-        child: 'td:nth-child(2)',
-        expectedText: 'Alice Martin'
-      });
-    });
-
-    await test.step('Verify presence by attribute (no assertion fields)', async () => {
-      await steps.verifyListedElement('TablePage', 'rows', {
-        attribute: { name: 'data-testid', value: 'table-row-1' }
-      });
-    });
-
-    log('TC_052 verifyListedElement — passed');
-  });
-});
-
-test.describe('TC_053: getListedElementData', () => {
-
-  test('all extraction variants', async ({ steps }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('Extract text by text match (no child)', async () => {
-      const text = await steps.getListedElementData('TablePage', 'rows', { text: 'Alice' });
-      expect(text).toContain('Alice');
-    });
-
-    await test.step('Extract text from child by text match (text + child)', async () => {
-      const name = await steps.getListedElementData('TablePage', 'rows', {
-        text: 'Alice',
-        child: 'td:nth-child(2)'
-      });
-      expect(name).toBe('Alice Martin');
-    });
-
-    await test.step('Extract attribute by text match (text + extractAttribute)', async () => {
-      const testId = await steps.getListedElementData('TablePage', 'rows', {
-        text: 'Alice',
-        extractAttribute: 'data-testid'
-      });
-      expect(testId).toBe('table-row-1');
-    });
-
-    await test.step('Extract text by attribute match (attribute only)', async () => {
-      const text = await steps.getListedElementData('TablePage', 'rows', {
-        attribute: { name: 'data-testid', value: 'table-row-1' }
-      });
-      expect(text).toContain('Alice');
-    });
-
-    await test.step('Extract attribute from child by attribute match (attribute + child + extractAttribute)', async () => {
-      await steps.navigateTo('/');
-      const testId = await steps.getListedElementData('HomePage', 'categories', {
-        text: 'Forms',
-        extractAttribute: 'data-testid'
-      });
-      expect(testId).toBe('home-card-forms');
-    });
-
-    await test.step('Extract text from child by attribute match (attribute + child)', async () => {
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-      const name = await steps.getListedElementData('TablePage', 'rows', {
-        attribute: { name: 'data-testid', value: 'table-row-1' },
-        child: 'td:nth-child(2)'
-      });
-      expect(name).toBe('Alice Martin');
-    });
-
-    log('TC_053 getListedElementData — passed');
-  });
-});
-
-test.describe('TC_054: getListedElement (raw) — child variants and error cases', () => {
-
-  test('child page-repo reference and error handling', async ({ page, repo, steps, interactions }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('getListedElement with text match', async () => {
-      const baseLocator = page.locator(repo.getSelector('TablePage', 'rows'));
-      const row = await interactions.interact.getListedElement(baseLocator, { text: 'Alice' }, repo);
-      const text = await row.textContent();
-      expect(text).toContain('Alice');
-    });
-
-    await test.step('getListedElement with attribute match', async () => {
-      const baseLocator = page.locator(repo.getSelector('TablePage', 'rows'));
-      const row = await interactions.interact.getListedElement(baseLocator, {
-        attribute: { name: 'data-testid', value: 'table-row-1' }
-      }, repo);
-      const text = await row.textContent();
-      expect(text).toContain('Alice');
-    });
-
-    await test.step('getListedElement with child as page-repo reference', async () => {
-      const baseLocator = page.locator(repo.getSelector('TablePage', 'rows'));
-      const child = await interactions.interact.getListedElement(baseLocator, {
-        text: 'Alice',
-        child: { pageName: 'TablePage', elementName: 'rowCheckboxes' }
-      }, repo);
-      expect(child).toBeTruthy();
-    });
-
-    await test.step('getListedElement throws when child is page-repo ref but no repo provided', async () => {
-      const baseLocator = page.locator(repo.getSelector('TablePage', 'rows'));
-      let errorThrown = false;
-      try {
-        await interactions.interact.getListedElement(baseLocator, {
-          text: 'Alice',
-          child: { pageName: 'TablePage', elementName: 'rowCheckboxes' }
-        });
-      } catch (e: unknown) {
-        errorThrown = true;
-        expect((e as Error).message).toContain('ElementRepository instance is required');
-      }
-      expect(errorThrown).toBe(true);
-    });
-
-    await test.step('getListedElement throws when neither text nor attribute provided', async () => {
-      const baseLocator = page.locator(repo.getSelector('TablePage', 'rows'));
-      let errorThrown = false;
-      try {
-        await interactions.interact.getListedElement(baseLocator, {}, repo);
-      } catch (e: unknown) {
-        errorThrown = true;
-        expect((e as Error).message).toContain('requires either "text" or "attribute"');
-      }
-      expect(errorThrown).toBe(true);
-    });
-
-    log('TC_054 Raw getListedElement — passed');
-  });
-});
-
-// ──────────────────────────────────────────────────────────────────────────────
-// New Methods — TC_055 through TC_066
-// ──────────────────────────────────────────────────────────────────────────────
-
-test.describe('TC_055: fillForm — fill multiple fields in one call', () => {
-
-  test('fills text inputs and selects a dropdown', async ({ steps }) => {
-
-    await test.step('Navigate to Forms page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('HomePage', 'formsCard');
-      await steps.verifyUrlContains('/forms');
-    });
-
-    await test.step('Fill form with text fields and dropdown', async () => {
-      await steps.fillForm('FormsPage', {
-        nameInput: 'Fill Form Test',
-        emailInput: 'fillform@test.com',
-        mobileInput: '1234567890',
-        genderDropdown: { type: DropdownSelectType.VALUE, value: 'Male' }
-      });
-    });
-
-    await test.step('Verify all fields were filled', async () => {
-      await steps.verifyInputValue('FormsPage', 'nameInput', 'Fill Form Test');
-      await steps.verifyInputValue('FormsPage', 'emailInput', 'fillform@test.com');
-      await steps.verifyInputValue('FormsPage', 'mobileInput', '1234567890');
-    });
-
-    log('TC_055 fillForm — passed');
-  });
-});
-
-test.describe('TC_056: clearInput — clear an input field', () => {
-
-  test('fills then clears an input', async ({ steps }) => {
-
-    await test.step('Navigate to Text Inputs page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('Fill text input', async () => {
-      await steps.fill('TextInputsPage', 'textInput', 'Some text to clear');
-      await steps.verifyInputValue('TextInputsPage', 'textInput', 'Some text to clear');
-    });
-
-    await test.step('Clear input and verify it is empty', async () => {
-      await steps.clearInput('TextInputsPage', 'textInput');
-      await steps.verifyInputValue('TextInputsPage', 'textInput', '');
-    });
-
-    log('TC_056 clearInput — passed');
-  });
-});
-
-test.describe('TC_057: getInputValue — read input values', () => {
-
-  test('reads pre-populated and user-entered values', async ({ steps }) => {
-
-    await test.step('Navigate to Text Inputs page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('Read pre-populated error input value', async () => {
-      const value = await steps.getInputValue('TextInputsPage', 'errorInput');
-      expect(value).toBe('invalid@');
-    });
-
-    await test.step('Read pre-populated success input value', async () => {
-      const value = await steps.getInputValue('TextInputsPage', 'successInput');
-      expect(value).toBe('valid@example.com');
-    });
-
-    await test.step('Fill and read back a value', async () => {
-      await steps.fill('TextInputsPage', 'emailInput', 'test@example.org');
-      const value = await steps.getInputValue('TextInputsPage', 'emailInput');
-      expect(value).toBe('test@example.org');
-    });
-
-    log('TC_057 getInputValue — passed');
-  });
-});
-
-test.describe('TC_058: getCount — count elements', () => {
-
-  test('counts categories on home page', async ({ steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('getCount returns 8 categories', async () => {
-      const count = await steps.getCount('HomePage', 'categories');
-      expect(count).toBe(8);
-    });
-
-    log('TC_058 getCount — passed');
-  });
-});
-
-test.describe('TC_059: getAll — bulk text extraction', () => {
-
-  test('extracts texts and attributes from table rows', async ({ steps }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('getAll — basic text extraction from rows', async () => {
-      const texts = await steps.getAll('TablePage', 'rows');
-      expect(texts.length).toBe(5);
-      expect(texts[0]).toContain('Alice Martin');
-    });
-
-    await test.step('getAll — with child selector', async () => {
-      const names = await steps.getAll('TablePage', 'rows', { child: { pageName: 'TablePage', elementName: 'nameCell' } });
-      expect(names).toEqual(['Alice Martin', 'Bob Chen', 'Carol White', 'David Kim', 'Eve Torres']);
-    });
-
-    await test.step('getAll — with extractAttribute', async () => {
-      const testIds = await steps.getAll('TablePage', 'rowCheckboxes', { extractAttribute: 'data-testid' });
-      expect(testIds.length).toBe(5);
-      expect(testIds[0]).toContain('table-row-checkbox-');
-    });
-
-    log('TC_059 getAll — passed');
-  });
-});
-
-test.describe('TC_060: getCssProperty — read computed styles', () => {
-
-  test('reads CSS properties from elements', async ({ steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('Read display property of page title', async () => {
-      const display = await steps.getCssProperty('HomePage', 'pageTitle', 'display');
-      expect(display).toBeTruthy();
-    });
-
-    await test.step('Read font-size of page title', async () => {
-      const fontSize = await steps.getCssProperty('HomePage', 'pageTitle', 'font-size');
-      expect(fontSize).toMatch(/\d+px/);
-    });
-
-    log('TC_060 getCssProperty — passed');
-  });
-});
-
-test.describe('TC_061: verifyCssProperty — assert computed styles', () => {
-
-  test('asserts CSS property values', async ({ steps }) => {
-
-    await test.step('Navigate to Text Inputs page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'textInputsLink');
-      await steps.verifyUrlContains('/text-inputs');
-    });
-
-    await test.step('Verify display property of text input', async () => {
-      const display = await steps.getCssProperty('TextInputsPage', 'textInput', 'display');
-      await steps.verifyCssProperty('TextInputsPage', 'textInput', 'display', display);
-    });
-
-    log('TC_061 verifyCssProperty — passed');
-  });
-});
-
-test.describe('TC_062: waitAndClick — wait then click', () => {
-
-  test('waits for visible state then clicks', async ({ steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('waitAndClick on forms card (default: visible)', async () => {
-      await steps.waitAndClick('HomePage', 'formsCard');
-      await steps.verifyUrlContains('/forms');
-    });
-
-    log('TC_062 waitAndClick — passed');
-  });
-});
-
-test.describe('TC_063: clickNth — click by index', () => {
-
-  test('clicks an element at a specific index', async ({ steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('Click the second category card (index 1)', async () => {
-      await steps.clickNth('HomePage', 'categories', 1);
-      // Should navigate to one of the category pages
-      await steps.verifyAbsence('HomePage', 'categories');
-    });
-
-    log('TC_063 clickNth — passed');
-  });
-});
-
-test.describe('TC_064: selectMultiple — select multiple options', () => {
-
-  test('selects multiple values in a multi-select', async ({ steps }) => {
-
-    await test.step('Navigate to Dropdown page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'dropdownLink');
-      await steps.verifyUrlContains('/dropdown');
-    });
-
-    await test.step('Select multiple countries', async () => {
-      const selected = await steps.selectMultiple('DropdownSelectPage', 'multiSelect', ['Australia', 'Canada', 'France']);
-      expect(selected).toContain('Australia');
-      expect(selected).toContain('Canada');
-      expect(selected).toContain('France');
-    });
-
-    log('TC_064 selectMultiple — passed');
-  });
-});
-
-test.describe('TC_065: verifyListOrder — assert sort direction', () => {
-
-  test('verifies ascending sort on table names', async ({ page, steps }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('Default name order is ascending — verify with verifyListOrder', async () => {
-      await steps.verifyListOrder('TablePage', 'nameCell', 'asc');
-    });
-
-    await test.step('Click Name header twice to sort descending', async () => {
-      await steps.click('TablePage', 'headerName');
-      await steps.click('TablePage', 'headerName');
-    });
-
-    await test.step('Verify descending order with verifyListOrder', async () => {
-      await steps.verifyListOrder('TablePage', 'nameCell', 'desc');
-    });
-
-    log('TC_065 verifyListOrder — passed');
-  });
-});
-
-test.describe('TC_066: retryUntil — retry action until verification', () => {
-
-  test('retries clicking until expected state', async ({ steps }) => {
-
-    await test.step('Navigate to Checkboxes page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'checkboxesLink');
-      await steps.verifyUrlContains('/checkboxes');
-    });
-
-    await test.step('retryUntil — click checkbox until checked', async () => {
-      // First uncheck if already checked
-      await steps.uncheck('CheckboxesPage', 'uncheckedCheckbox');
-      // Now retry click until it becomes checked
-      await steps.retryUntil(
-        async () => { await steps.check('CheckboxesPage', 'uncheckedCheckbox'); },
-        async () => { await steps.verifyState('CheckboxesPage', 'uncheckedCheckbox', 'checked'); },
-        3,
-        500
-      );
-    });
-
-    log('TC_066 retryUntil — passed');
-  });
-});
-
-test.describe('TC_067: screenshot — page and element screenshots', () => {
-
-  test('captures page and element screenshots', async ({ steps }) => {
-
-    await test.step('Navigate to home page', async () => {
-      await steps.navigateTo('/');
-    });
-
-    await test.step('Page screenshot (no options)', async () => {
-      const buffer = await steps.screenshot();
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(buffer.length).toBeGreaterThan(0);
-    });
-
-    await test.step('Page screenshot with fullPage option', async () => {
-      const buffer = await steps.screenshot({ fullPage: true });
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(buffer.length).toBeGreaterThan(0);
-    });
-
-    await test.step('Element screenshot', async () => {
-      const buffer = await steps.screenshot('HomePage', 'pageTitle');
-      expect(buffer).toBeInstanceOf(Buffer);
-      expect(buffer.length).toBeGreaterThan(0);
-    });
-
-    log('TC_067 screenshot — passed');
-  });
-});
-
-test.describe('TC_068: waitForNetworkIdle — wait for network quiet', () => {
-
-  test('waits for network idle after navigation', async ({ steps }) => {
-
-    await test.step('Navigate and wait for network idle', async () => {
-      await steps.navigateTo('/');
-      await steps.waitForNetworkIdle();
-    });
-
-    await test.step('Page is fully loaded after network idle', async () => {
-      await steps.verifyPresence('HomePage', 'pageTitle');
-      await steps.verifyCount('HomePage', 'categories', { exactly: 8 });
-    });
-
-    log('TC_068 waitForNetworkIdle — passed');
-  });
-});
-
-test.describe('TC_069: verifyOrder — assert exact element text order', () => {
-
-  test('verifies exact text order of table name cells', async ({ steps }) => {
-
-    await test.step('Navigate to Table page', async () => {
-      await steps.navigateTo('/');
-      await steps.click('SidebarNav', 'tableLink');
-      await steps.verifyUrlContains('/table');
-    });
-
-    await test.step('Verify exact order of names using verifyOrder', async () => {
-      await steps.verifyOrder('TablePage', 'nameCell', ['Alice Martin', 'Bob Chen', 'Carol White', 'David Kim', 'Eve Torres']);
-    });
-
-    log('TC_069 verifyOrder — passed');
-  });
-});
